+++ MeetingBar/Views/Changelog/Changelog.swift
@@ -95,7 +95,7 @@ struct ChangelogView: View {
                 }
                 if compareVersions("3.11.0", lastRevisedVersionInChangelog) {
                     Section(header: Text("Version 3.11.0")) {
-                        Text("• Added integration with Pop, Livestorm, Chorus & Gong")
+                        Text("• Added integration with Pop, Livestorm, Chorus, Gong & Yandex.Telemost")
                         Text("• Fix readability of the statusbar text in multi-screen setups (#354)")
                         Text("• Detect hidden menubar icon (#429)")
                         Text("😴 Added feature to snooze the notification (#)")