+++ MeetingBar/EventStores/EKEventStore.swift
@@ -8,6 +8,13 @@
 import EventKit
 import PromiseKit
 
+// Ref: https://stackoverflow.com/a/66074963
+extension EKParticipant {
+    var safeNSURL: NSURL? {
+        perform(#selector(getter: EKParticipant.url))?.takeUnretainedValue() as? NSURL
+    }
+}
+
 extension EKEventStore: EventStore {
     static let shared = initEKEventStore()
 
@@ -102,7 +109,8 @@ extension EKEventStore: EventStore {
                     }
 
                     let optional = rawAttendee.participantRole == .optional
-                    let email = String(describing: rawAttendee.url).replacingOccurrences(of: "mailto:", with: "")
+                    let email = rawAttendee.safeNSURL?.resourceSpecifier
+                    print(email)
                     let attendee = MBEventAttendee(email: email, name: rawAttendee.name, status: attendeeStatus, optional: optional, isCurrentUser: rawAttendee.isCurrentUser)
 
                     attendees.append(attendee)